pyproject.toml
[project]
name = "elastic-claude-agent-skill-demo"
version = "0.1.0"
description = "Elasticsearch backend for Claude agent skills demo"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "elasticsearch>=8.0.0",
    "python-dotenv>=1.0.0",
    "pytest>=7.0.0",
    "pytest-json-report>=1.5.0",
    "requests>=2.31.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["scripts", "tests", "sample_skills"]
